Conflicts:
	libraries/RC_Channel/RC_Channel.cpp
	libraries/RC_Channel/RC_Channel.h

+++ Tools/scripts/install-prereqs-arch.sh
@@ -1,24 +1,22 @@
 #!/bin/bash
 set -e
-
-command -v yaourt >/dev/null 2>&1 || { echo >&2 "Please install yaourt first. Aborting."; exit 1; }
+set -x
 
 CWD=$(pwd)
 OPT="/opt"
 
-BASE_PKGS="wget curl base-devel git-core tk gsfonts"
-SITL_PKGS="python2-pip python-pip wxpython2.8 opencv python2-numpy python2-scipy ccache"
-PX4_PKGS="lib32-glibc zip zlib ncurses"
+BASE_PKGS="base-devel ccache git gsfonts tk wget"
+SITL_PKGS="python2-pip python-pip wxpython opencv python2-numpy python2-scipy"
+PX4_PKGS="lib32-glibc zip zlib ncurses cmake"
 
-PYTHON2_PKGS="pymavlink MAVProxy droneapi argparse matplotlib pyparsing catkin_pkg"
+PYTHON2_PKGS="future lxml pymavlink MAVProxy argparse matplotlib pyparsing"
 PYTHON3_PKGS="pyserial empy"
-ARCH_AUR_PKGS="genromfs"
 
 # GNU Tools for ARM Embedded Processors
 # (see https://launchpad.net/gcc-arm-embedded/)
-ARM_ROOT="gcc-arm-none-eabi-4_7-2014q2"
-ARM_TARBALL="$ARM_ROOT-20140408-linux.tar.bz2"
-ARM_TARBALL_URL="http://firmware.diydrones.com/Tools/PX4-tools/$ARM_TARBALL"
+ARM_ROOT="gcc-arm-none-eabi-6-2017-q2-update"
+ARM_TARBALL="$ARM_ROOT-linux.tar.bz2"
+ARM_TARBALL_URL="http://firmware.ardupilot.org/Tools/STM32-tools/$ARM_TARBALL"
 
 # Ardupilot Tools
 ARDUPILOT_TOOLS="ardupilot/Tools/autotest"
@@ -34,15 +32,18 @@ function prompt_user() {
 
 sudo usermod -a -G uucp $USER
 
-sudo pacman -S --noconfirm $BASE_PKGS $SITL_PKGS $PX4_PKGS
-sudo pip2 -q install $PYTHON2_PKGS
-sudo pip3 -q install $PYTHON3_PKGS
-yaourt -S --noconfirm $ARCH_AUR_PKGS
+sudo pacman -Sy --noconfirm --needed $BASE_PKGS $SITL_PKGS $PX4_PKGS
+pip2 -q install --user -U $PYTHON2_PKGS
+pip3 -q install --user -U $PYTHON3_PKGS
 
 (
- cd /usr/lib/ccache
- sudo ln -s /usr/bin/ccache arm-none-eabi-g++
- sudo ln -s /usr/bin/ccache arm-none-eabi-gcc
+    cd /usr/lib/ccache
+    if [ ! -f arm-none-eabi-g++ ]; then
+       sudo ln -s /usr/bin/ccache arm-none-eabi-g++
+    fi
+    if [ ! -f arm-none-eabi-g++ ]; then
+        sudo ln -s /usr/bin/ccache arm-none-eabi-gcc
+    fi
 )
 
 if [ ! -d $OPT/$ARM_ROOT ]; then
@@ -56,7 +57,7 @@ fi
 
 exportline="export PATH=$OPT/$ARM_ROOT/bin:\$PATH";
 if ! grep -Fxq "$exportline" ~/.bashrc ; then
-    if prompt_user "Add $OPT/$ARM_ROOT/bin to your PATH [Y/n]?" ; then
+    if prompt_user "Add $OPT/$ARM_ROOT/bin to your PATH [N/y]?" ; then
         echo "$exportline" >> ~/.bashrc
         . ~/.bashrc
     else
@@ -66,7 +67,7 @@ fi
 
 exportline2="export PATH=$CWD/$ARDUPILOT_TOOLS:\$PATH";
 if  ! grep -Fxq "$exportline2" ~/.bashrc ; then
-    if prompt_user "Add $CWD/$ARDUPILOT_TOOLS to your PATH [Y/n]?" ; then
+    if prompt_user "Add $CWD/$ARDUPILOT_TOOLS to your PATH [N/y]?" ; then
         echo "$exportline2" >> ~/.bashrc
         . ~/.bashrc
     else
@@ -74,10 +75,10 @@ if  ! grep -Fxq "$exportline2" ~/.bashrc ; then
     fi
 fi
 
+SCRIPT_DIR=$(dirname $(realpath ${BASH_SOURCE[0]}))
 (
- cd ./ardupilot
- git submodule init
- git submodule update
+    cd $SCRIPT_DIR
+    git submodule update --init --recursive
 )
 
 echo "Done. Please log out and log in again."
